% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{add_labs}
\alias{add_labs}
\title{Helper function to quickly add title, subtitle and caption or source}
\usage{
add_labs(gt_object, title = NULL, subtitle = NULL, caption = NULL, ...)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{title}{Text passed to the \code{tab_header()} function}

\item{subtitle}{Text passed to the \code{tab_header()} function}

\item{caption}{Text passed to the \code{tab_source_note()} function}

\item{...}{Other gt options}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Quick way to add a table title, subtitle and source.
}
