% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{adjust_row_padding}
\alias{adjust_row_padding}
\title{adjust_row_padding}
\usage{
adjust_row_padding(gt_object, padding_setting = "regular")
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{padding_setting}{Adjusts the density of the text in a table}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
adjust_row_padding
}
