% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{align_columns}
\alias{align_columns}
\title{Align columns}
\usage{
align_columns(gt_object)
}
\arguments{
\item{gt_object}{}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Aligns columns based on their type. Numeric columns are right aligned.
Text columns are left aligned.
}
