% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{apply_row_striping}
\alias{apply_row_striping}
\title{Row striping}
\usage{
apply_row_striping(gt_object, stripe_color = "#E8E8E9")
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{stripe_color}{Fill color for the stripes, defaults to \code{glitr::si_palettes$hw_slate_t[5]}}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Add faint row striping to a table based on your color of choice.
}
