% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{check_gt_object}
\alias{check_gt_object}
\title{check_gt_object}
\usage{
check_gt_object(gt_object)
}
\arguments{
\item{gt_object}{An object of class \code{gt_tbl} to be checked.}
}
\value{
an error message
}
\description{
check_gt_object
}
