% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numeric_columns.R
\name{format_numeric_columns}
\alias{format_numeric_columns}
\title{Formats numeric columns}
\usage{
format_numeric_columns(gt_object)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Provides pre-set formatting for whole numbers and decimals.
All whole numbers are formatted with a comma, all decimals with a percentage.
}
