% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_column_as_image.R
\name{generate_achv_html}
\alias{generate_achv_html}
\title{Generate Achievement HTML}
\usage{
generate_achv_html(filenames)
}
\arguments{
\item{filenames}{The name of the image file. Acceptable inputs are \code{above_target}, \code{at_risk}, \code{concerned}, \code{on_target} and \code{achv_legend}.}
}
\value{
The URL string to the image.
}
\description{
Generates the HTML for embedding an image corresponding to the achievement status.
Acceptable inputs are: \code{above_target}, \code{at_risk}, \code{concerned}, and \code{on_target} and \code{achv_legend}.
}
\examples{
generate_achv_html("at_risk")
}
