% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_column_as_image.R
\name{generate_achv_url}
\alias{generate_achv_url}
\title{Generate Achievement URL}
\usage{
generate_achv_url(filename, url_suffix = ".png")
}
\arguments{
\item{filename}{The name of the image file.}

\item{url_suffix}{The ending of the file. Can be .png or .svg}
}
\value{
A complete URL string to the image.
}
\description{
Creates a URL to one of four images stored with the SI gt adornment package.
Options are: \code{above_target}, \code{at_risk}, \code{concerned}, and \code{ontarget}.
}
\examples{
generate_achv_url("above_target")
generate_achv_url("at_risk")
}
