% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_column_as_image.R
\name{render_column_as_image}
\alias{render_column_as_image}
\title{Render Column as Image}
\usage{
render_column_as_image(gt_object, columns, height = 30)
}
\arguments{
\item{gt_object}{A gt table object.}

\item{columns}{The columns to transform into images.}

\item{height}{The height of the images in pixels (default is 30).}
}
\value{
A gt table object with specified columns rendered as images.
}
\description{
Transforms specified columns in a gt table to display images instead of text. The images are specified by their URLs.
}
