% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_gt_base.R
\name{si_gt_base}
\alias{si_gt_base}
\title{SI base theme}
\usage{
si_gt_base(
  gt_object,
  header_fill = "#626672",
  header_font_color = "white",
  table_font = "Source Sans Pro",
  table_font_color = "#626672",
  ...
)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{header_fill}{Fill color for the header background, defaults to \code{glitr::hw_slate}}

\item{header_font_color}{Font color for the header text, defaults to white}

\item{table_font}{Font type as called from \code{google_font()}}

\item{table_font_color}{Font color for the table body, defaults to \code{glitr::hw_slate}}

\item{...}{Optional additional arguments to \code{gt::table_options()}}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
SI base theme
}
